App compiles but rust cant find audio_analyzer.py for some reason.
Implemented files and functions:
audio_analyzer.py
-analyze

audio_analysis.rs
-analyze_audio
-my_rust_module

main.rs
-analyze_and_update_beat
-analyze_audio_command

Also in main.rs
Added a couple lines to main to setup python.
Added a call to analyze_and_update_beat at the end of add_beat function.

Also cargo.toml has updates for pyo3 functionality.

To run everything you will need to install requirements.txt into a venv and run the venv when running the app. If requirements.txt doesnt work on install, just install librosa library, since i just installed librosa before freezing requirements into requirements.txt

+++ src-tauri/src/models.rs
@@ -32,7 +32,7 @@ pub struct Beat {
     pub cover_art: Option<String>,
     pub comments: Option<String>,
     pub file_path: String,
-    pub bpm: Option<i32>,
+    pub bpm: Option<f64>,
     pub musical_key: Option<String>,
     pub date_created: NaiveDateTime,
 }
@@ -52,7 +52,7 @@ pub struct NewBeat<'a> {
     pub cover_art: Option<&'a str>,
     pub comments: Option<&'a str>,
     pub file_path: &'a str,
-    pub bpm: Option<i32>,
+    pub bpm: Option<f64>,
     pub musical_key: Option<&'a str>,
     pub date_created: &'a str,
 }
